feature: Adding JSON extract function, ListView and Adapter
We construct a QueryUtils class to extract JSON data from string. We also added a vector Drawable resource instead of bitmap resources for UI. Finally we displayed the test data on a ListView throught an adapter.

	-new file:   app/src/main/res/drawable/section_drawable.xml
This is a file where we defined a vector drawable resource on a xml to draw a newspaper on the background to display an incon of the inflated views of the ListView
	-new file:   app/src/main/res/layout/list_item.xml
	Ths file is added to contain the information retraived from each item of the news api. This is the inflated view for an specific new

Issue #001: list_item should contemplate larger words on the news_icon. The text is overlapng the newspaper icon.